+++ apps/forms/components/formfield/file-upload/index.ts
@@ -1,1 +1,2 @@
 export * from "./file-upload";
+export * from "./uploader";
